+++ packages/polaris-viz/src/components/StackedAreaChart/components/Area/index.ts
@@ -1,1 +1,2 @@
 export {Area} from './Area';
+export {AnimatedArea} from './AnimatedArea';